[Gold V] Title: 가로등, Time: 364 ms, Memory: 62804 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/32069. 가로등/가로등.cc
@@ -6,7 +6,7 @@ using namespace std;
 #define INF 987654321
 ll n, l, k;
 vector<ll> v;
-int dx[] = {1, -1};
+ll dx[] = {1, -1};
 void bfs()
 {
     queue<ll> q;
@@ -30,7 +30,7 @@ void bfs()
 
             for (int i = 0; i < 2; i++)
             {
-                int next = now + dx[i];
+                ll next = now + dx[i];
 
                 if (next < 0 || next > l)
                     continue;
